Add physics service

+++ src/treasure-hunter/Code/Infrastructure/Physics/ColliderToEntityRegistryResolver.cs
@@ -5,13 +5,13 @@ namespace Code.Infrastructure.Physics
 {
   class ColliderToEntityRegistryResolver : IColliderToEntityRegistry, IEntityResolverFromCollider
   {
-    private readonly Dictionary<int, IEntity> _entityByInstanceId = new();
+    private readonly Dictionary<ulong, IEntity> _entityByInstanceId = new();
     
-    public void Register(int instanceId, IEntity entity) => _entityByInstanceId[instanceId] = entity;
+    public void Register(ulong instanceId, IEntity entity) => _entityByInstanceId[instanceId] = entity;
 
-    public void Unregister(int instanceId) => _entityByInstanceId.Remove(instanceId);
+    public void Unregister(ulong instanceId) => _entityByInstanceId.Remove(instanceId);
 
-    public TEntity Resolve<TEntity>(int instanceId) where TEntity : class =>
+    public TEntity Resolve<TEntity>(ulong instanceId) where TEntity : class =>
       _entityByInstanceId.TryGetValue(instanceId, out IEntity entity)
         ? (TEntity)entity
         : null;
